requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
altair==5.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
basic_pitch_torch @ git+https://github.com/sen-sourav/basic-pitch-torch.git@efe97965e21608a8284b2ae215a631e474454187
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
discord==2.3.2
discord.py==2.4.0
einops==0.8.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mido==1.3.2
mir_eval==0.7
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
narwhals==1.3.0
networkx==3.1
nnAudio==0.3.3
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pooch==1.8.2
pretty-midi==0.2.10
protobuf==5.27.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
resampy==0.4.3
responses==0.25.3
rich==13.7.1
rpds-py==0.20.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
stqdm==0.0.5
streamlit==1.37.1
streamlit-audiorecorder==0.0.5
streamlit-js-eval==0.1.7
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.4.0
torch-summary==1.4.5
torchaudio==2.4.0
torchsummary==1.5.1
tornado==6.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultimate_accompaniment_transformer @ git+https://github.com/sen-sourav/Ultimate-Accompaniment-Transformer.git@7eb255c4253ea7fce6f36ecf16961dce6cf0423a
urllib3==2.2.2
watchdog==4.0.2
yarl==1.9.4
zipp==3.19.2
pyfluidsynth
